chore: move all validation to a separate file
